+++ src/controllers/user/index.ts
@@ -1,0 +1,5 @@
+import { create } from './create';
+
+export default {
+  create,
+};
