+++ src/models/user/index.ts
@@ -1,2 +1,2 @@
-export { default as User } from './user';
+export { default as User, UserInstance } from './user';
 export { generateKey, generateHash, compareHash } from './utils';
